+++ CHANGELOG.md
@@ -14,6 +14,9 @@
     - Use `static_id!` to declare static IDs, the new way is compatible with hot reloading.
     - Removed `StaticWindowId`, `StaticMonitorId`, `StaticPropertyId`, `StaticAppId`, `StaticWidgetId`, `StaticDeviceId`, `StaticStateId`,  `StaticCommandMetaVarId`, `StaticSpatialFrameId`.
 * Implemented equality/hash for `zng::task::SignalOnce`.
+* Breaking `Window::on_close` now has args type `WindowCloseArgs`.
+    - Add `on_pre_window_close` and `on_window_close`.
+* Fix headless windows not receiving close events on app exit request.
 * Add `std::env::consts::OS` to crash error.
 
 [`zng.code-snippets`]: .vscode/zng.code-snippets